requirements.txt
click
langdetect
nltk
numpy
pandas
pyspellchecker
spacy
tqdm